+++ src/app/common/user/user-state.model.ts
@@ -1,5 +1,0 @@
-import {UserBusinessModel} from "../business-models/user.business-model";
-
-export interface UserStateModel {
-  user: UserBusinessModel | null;
-}