+++ src/app/features/login/services/login.service.ts
@@ -11,7 +11,7 @@ export const TOKEN_KEY = 'auth.token'
 @Injectable({
   providedIn: 'root'
 })
-export class AuthService {
+export class LoginService {
 
   constructor(private http: HttpClient) {
   }